增加shiro多个realm配置

+++ src/main/java/com/example/music/Controller/UserController.java
@@ -7,6 +7,8 @@ import com.example.music.VO.UserInfo;
 import org.apache.shiro.SecurityUtils;
 import org.apache.shiro.authc.AuthenticationException;
 import org.apache.shiro.authc.UsernamePasswordToken;
+import org.apache.shiro.authz.annotation.RequiresPermissions;
+import org.apache.shiro.authz.annotation.RequiresRoles;
 import org.apache.shiro.crypto.SecureRandomNumberGenerator;
 import org.apache.shiro.crypto.hash.SimpleHash;
 import org.apache.shiro.subject.Subject;
@@ -15,6 +17,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.servlet.http.HttpServletRequest;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -70,12 +73,13 @@ public class UserController {
         User userQuery = userService.selectByUserName(username);
         Integer id = userQuery.getId();
         String avator = userQuery.getAvator();
+        Boolean isVip=userQuery.getIsVip();
         System.out.println(id);
         Subject subject = SecurityUtils.getSubject();
         UsernamePasswordToken usernamePasswordToken = new UsernamePasswordToken(username, password);
         try {
             subject.login(usernamePasswordToken);
-            return new Response(200, "success", new UserInfo(id,username,avator));
+            return new Response(200, "success", new UserInfo(id,username,avator,isVip));
         } catch (AuthenticationException e) {
             return new Response(500, "failure", null);
         }
@@ -98,6 +102,22 @@ public class UserController {
             }
         }else{
             return new Response(500,"更新失败",null);
+        }
+    }
+
+    @PostMapping("/changeIsVip")
+    public Response changeIsVip(HttpServletRequest request){
+        String isVip = request.getParameter("isVip");
+        String id = request.getParameter("id");
+        Boolean isVip1=false;
+        if(isVip.equals("true")){
+            isVip1=true;
+        }
+        boolean flag = userService.changeIsVip(Integer.parseInt(id), isVip1);
+        if(flag){
+            return new Response(200,"更新状态成功",null);
+        }else{
+            return new Response(500,"更新状态失败",null);
         }
     }
 
@@ -137,11 +157,16 @@ public class UserController {
      */
     @GetMapping("/deleteUser/{id}")
     public Response deleteUser(@PathVariable(value="id")int id){
-        boolean flag = userService.delete(id);
-        if(flag){
-            return new Response(200,"删除成功",null);
+        Subject subject = SecurityUtils.getSubject();
+        if(subject.hasRole("role_user_admin")) {
+            boolean flag = userService.delete(id);
+            if (flag) {
+                return new Response(200, "删除成功", null);
+            } else {
+                return new Response(500, "删除失败", null);
+            }
         }else{
-            return new Response(500,"删除失败",null);
+            return new Response(500, "删除失败", null);
         }
     }
 
@@ -161,9 +186,42 @@ public class UserController {
         }
     }
 
+    @GetMapping("/getAcator/{id}")
+    public String getAcator(@PathVariable("id")String id){
+        User user = userService.selectByPrimaryKey(Integer.parseInt(id));
+        String avator = user.getAvator();
+        return avator;
+    }
+
+    @PostMapping("/setPassword")
+    public Response setPassword(HttpServletRequest request){
+        String id = request.getParameter("id");
+        String oldPassword = request.getParameter("oldPassword");
+        String newPassword = request.getParameter("newPassword");
+        User user = userService.selectByPrimaryKey(Integer.parseInt(id));
+        String password = user.getPassword();
+        String salt = user.getSalt();
+        String hashAlgorithmName = "MD5";
+        int hashIterations = 2;
+        String result = new SimpleHash(hashAlgorithmName,oldPassword,salt,hashIterations).toString();
+        if(result.equals(password)){
+            String salt1=new SecureRandomNumberGenerator().nextBytes().toString();
+            int times=2;
+            String algorithm="md5";
+            String pwdAfterHash=new SimpleHash(algorithm,newPassword,salt1,times).toString();
+            user.setSalt(salt1);
+            user.setPassword(pwdAfterHash);
+            userService.update(user);
+            return new Response(200,"修改密码成功",null);
+        }else{
+            return new Response(500,"旧密码不正确",null);
+        }
+
+    }
+
 
     /**
-     * 更新歌手头像
+     * 更新用户头像
      */
     @PostMapping(value="/updateUserAcator")
     public Response updateUserAcator(@RequestParam("file") MultipartFile avatorFile, @RequestParam("id")int id){
@@ -189,7 +247,7 @@ public class UserController {
         if(avator!=null){
             String subUrl = avator.substring(13);
             String oldFilePath = System.getProperty("user.dir") + System.getProperty("file.separator") + "img"
-                    + System.getProperty("file.separator") + "singerPic" + System.getProperty("file.separator") + subUrl;
+                    + System.getProperty("file.separator") + "userPic" + System.getProperty("file.separator") + subUrl;
             File file1 = new File(oldFilePath);
             file1.delete();
         }
